% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geog_lookup.R
\docType{data}
\name{geog_lookup}
\alias{geog_lookup}
\title{geog_lookup}
\format{
An object of class \code{data.frame} with 511 rows and 2 columns.
}
\usage{
geog_lookup
}
\description{
lookup table called by 'get_pbdb'
}
\keyword{datasets}
