% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTS2020.R
\docType{data}
\name{GTS2020}
\alias{GTS2020}
\title{GTS2020}
\format{
An object of class \code{data.frame} with 1535 rows and 9 columns.
}
\usage{
GTS2020
}
\description{
lookup table called by 'get_pbdb'
}
\keyword{datasets}
