% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTS2020_changelog.R
\docType{data}
\name{GTS2020_changelog}
\alias{GTS2020_changelog}
\title{GTS2020_changelog}
\format{
An object of class \code{list} of length 1.
}
\usage{
GTS2020_changelog
}
\description{
changelog of periodic updates made to the GTS2020 table originally
published in this package. The purpose of this changelog is to allow
the user to assess how up-to-date the resource is and made any changes
themselves if needed.
}
\keyword{datasets}
