% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph.R
\name{tgraph}
\alias{tgraph}
\title{tgraph}
\usage{
tgraph(x, ranks = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A dataframe containing a set of hierarchical
relationships. The leftmost column contains the
elements which will form the highest rank, followed
rightwards by successive ranks}

\item{ranks}{If not NULL, a vector of column names of x,
given in rank order. This is useful if x contains
columns which are not rank relevant or if columns are
not in hierarchical order. If not supplied, the
column order in x is used directly and is assumed to
be in rank order}

\item{verbose}{A logical indicating whether the
progress of tgraph construction should be reported
to the console}
}
\value{
a tgraph object
}
\description{
Function to create a tgraph representation of a
hierarchically organised dataframe. This is the focal
object of the t* functions - the complete set of
hierarchical relationships between a set of elements
}
