% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_graph.R
\name{update_graph}
\alias{update_graph}
\title{update_graph}
\usage{
update_graph(x, del = NULL, add = NULL, changes = NULL)
}
\arguments{
\item{x}{a tgraph object to modify}

\item{del}{A vector of element names or numbers to delete}

\item{add}{An edge sequence of edges to add to the graph}

\item{changes}{Alternatively, the output of
assess_duplicates, containing proposed deletions and
additions}
}
\value{
An updated tgraph object
}
\description{
Function to update the structure of a graph, given
a set of modification as returned by assess_duplicates
}
