% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_itp.R
\name{add_itp}
\alias{add_itp}
\title{add_itp}
\usage{
add_itp(x, taxon, legend.pos = "topright")
}
\arguments{
\item{x}{The list output of @seealso threshold_ranges}

\item{taxon}{A character vector of length one, specifying
one of the taxon names in x to be plotted}

\item{legend.pos}{One of topleft, bottomleft, topright or
bottomright, or a vector of length two, giving the xy
coordinates of the legend. A convenience parameter so that
the plot detail can remain unobscured.}
}
\value{
NULL, the detected peaks
}
\description{
Function to add detected peaks using the output of
}
\seealso{
threshold_ranges. This function should be used
to add information to an existing plot from @seealso densify,
ensuring that the same taxon name is being used
}
