% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brachios.R
\docType{data}
\name{brachios}
\alias{brachios}
\title{brachios}
\format{
An object of class \code{data.frame} with 151473 rows and 10 columns.
}
\usage{
brachios
}
\description{
An example dataset of Palaeozoic brachiopods
downloaded from the Paleobiology Database.
}
\keyword{datasets}
