% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_duplicates.R
\name{find_duplicates}
\alias{find_duplicates}
\title{find_duplicates}
\usage{
find_duplicates(x, ranks = NULL)
}
\arguments{
\item{x}{A hierarchically organised dataframe}

\item{ranks}{The ranks in the dataframe in which
to check for elements with multiple higher
classifications. The top rank is ignored by default}
}
\value{
A dataframe of elements with multiple
higher classifications and their ranks
}
\description{
Function to detect and report elements with multiple
higher assigments in a hierarchically structured
dataframe
}
\examples{
# load dataset
data("brachios")
b_ranks <- c("phylum", "class", "order", "family", "genus")
# run function
flag <- find_duplicates(brachios, ranks = b_ranks)
}
