% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTS_2020.R
\docType{data}
\name{GTS_2020}
\alias{GTS_2020}
\title{GTS_2020}
\format{
An object of class \code{data.frame} with 1534 rows and 9 columns.
}
\usage{
GTS_2020
}
\description{
lookup table called by 'get_pbdb'
}
\keyword{datasets}
