% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTS_2020_changelog.R
\docType{data}
\name{GTS_2020_changelog}
\alias{GTS_2020_changelog}
\title{GTS_2020_changelog}
\format{
An object of class \code{data.frame} with 1 rows and 2 columns.
}
\usage{
GTS_2020_changelog
}
\description{
changelog of periodic updates made to the GTS2020 table originally
published in this package. The purpose of this changelog is to allow
the user to assess how up-to-date the resource is and made any changes
themselves if needed. A data.frame with date-wise rows of edits
}
\keyword{datasets}
