% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_fields.R
\docType{data}
\name{pbdb_fields}
\alias{pbdb_fields}
\title{geog_lookup}
\format{
An object of class \code{list} of length 31.
}
\usage{
pbdb_fields
}
\description{
lookup table called by 'get_pbdb'
}
\keyword{datasets}
