% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_kingdoms.R
\docType{data}
\name{pbdb_kingdoms}
\alias{pbdb_kingdoms}
\title{pbdb_kingdoms}
\format{
An object of class \code{list} of length 3.
}
\usage{
pbdb_kingdoms
}
\description{
lookup table called by 'get_pbdb'
}
\keyword{datasets}
