% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sep_code.R
\docType{data}
\name{sep_code}
\alias{sep_code}
\title{sep_code}
\format{
An object of class \code{data.frame} with 306 rows and 8 columns.
}
\source{
chronosphere (fetch), Sepkoski 2002
}
\usage{
data(sep_code)
}
\description{
Lookup table of chronostratigraphic stage abbreviations
used in the Sepkoski Compendium, with interval boundaries
updated to the GTS2020 standard
}
\keyword{datasets}
