% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepkoski.R
\docType{data}
\name{sepkoski}
\alias{sepkoski}
\title{sepkoski}
\format{
An object of class \code{data.frame} with 35700 rows and 6 columns.
}
\usage{
sepkoski
}
\description{
An example dataset. A port of the Sepkoski Compendium
from the chronosphere package, with a few corrections
and GTS2020 dating applied
}
\keyword{datasets}
